latest

+++ code/arg-class.py
@@ -4,56 +4,50 @@ import subprocess
 import loader
 import os
 import json
-from tqdm import tqdm
 import numpy as np
 import shutil
+import write_out
 
+job_name = "kaggle_2000"
+path = "am/MARGOT/kaggle_2000"
+tool = "am/MARGOT/run_margot.sh" # Default AM tool
 
 def df_to_doc(clean_text):
     global counter
-    with open(f"am/MARGOT/version_1/temp/news/{counter}.txt", "a+") as text_file:
+    with open(f"{path}/temp/news/{counter}.txt", "a+") as text_file:
         text_file.write(clean_text)
     counter += 1
     
 def extract_argumentation():
     global step
-    tool = "am/MARGOT/run_margot.sh" # Default AM tool
-    for idx, file in enumerate(sorted(os.listdir("am/MARGOT/version_1/temp/news"), key=lambda x: int(x.split('.')[0]))):
+    for idx, file in enumerate(sorted(os.listdir(f"{path}/temp/news"), key=lambda x: int(x.split('.')[0]))):
 
-        print(idx, file)
-        total = len(os.listdir("am/MARGOT/version_1/temp/news"))
-        print("------------------------------------")
-        print(f"At step:{step}")
-        print("------------------------------------")
+        total = len(os.listdir(f"{path}/temp/news"))
+        print(f"STEP: {step}")
+        print("------------------------------------------------------------------------")
 
 
-        input = f"version_1/temp/news/{file}"
-        output = f"version_1/temp/arguments/{idx}"
+        input = f"{job_name}/temp/news/{file}"
+        output = f"{job_name}/temp/arguments/{idx}"
         subprocess.call([tool, input, output])
 
         step += 1
-        print()
 
         # Write out temporary results
-        if step in np.arange(0, total, 150):
-            print("Writing out temp results")
+        if step in np.arange(0, total, 50):
+            print("WRITE OUT TEMPORARY RESULTS")
+            print("------------------------------------------------------------------------")
             res = parse_output()
             temp_result = pd.DataFrame.from_dict(res, orient='index')
-            temp_result.to_csv(f"am/MARGOT/version_1/temp_results/temp_{name}.csv")
-            temp_result.to_csv(f"am/MARGOT/version_1/temp_results/temp_{name}.xlsx")            
-
+            write_out.write_data(temp_result, f"{path}/temp_results/temp_{name}")
 
     res = parse_output()
-
-    #res = {k: [v] for k, v in res.items()}
-
     return pd.DataFrame.from_dict(res, orient='index')
-
 
 def parse_output():
     argument_structures = {}
-    for dir in sorted(os.listdir("am/MARGOT/version_1/temp/arguments"), key=lambda x: int(x.split('.')[0])):
-        with open(f"am/MARGOT/version_1/temp/arguments/{dir}/OUTPUT.json") as json_file:
+    for dir in sorted(os.listdir(f"{path}/temp/arguments"), key=lambda x: int(x.split('.')[0])):
+        with open(f"{path}/temp/arguments/{dir}/OUTPUT.json") as json_file:
 
             doc = json.loads(json_file.read())
             args = {"claim": [], "evidence": []}
@@ -73,31 +67,31 @@ def parse_output():
 if __name__ == "__main__":
     # TODO: Implement command-line arguments so that different AM-tools can be used
     # parse_commands()
-    tqdm.pandas()
 
-    data = loader.load_data(arg=True)
+    data = loader.load_data_arg("data/original", "data/clean/arg")
     
     for name, df in data.items():
-        step = 0
-        counter = 0
-        print(f"Data: {name}, {len(df)}")
+        if name == "kaggle": ##
+            step = 0
+            counter = 0
 
-        df.dropna(inplace=True)
+            df.dropna(inplace=True)
 
-        dir = "am/MARGOT/version_1/temp"
-        if os.path.exists(dir):
-            shutil.rmtree(dir)
-        os.makedirs(f"{dir}/news")
-        os.makedirs(f"{dir}/arguments")
+            dir = f"{path}/temp"
+            if os.path.exists(dir):
+                shutil.rmtree(dir)
+            os.makedirs(f"{dir}/news")
+            os.makedirs(f"{dir}/arguments")
 
-        # Convert news values out to documents
-        df.progress_apply(lambda x: df_to_doc(x["text"]), axis=1)
-        # Run Margot over the documents
-        result = extract_argumentation()
+            # Convert news values out to documents
+            df.apply(lambda x: df_to_doc(x["text"]), axis=1)
 
-        # Write result data/argumentation_structure
-        result.to_csv(f"am/MARGOT/version_1/fin_{name}.csv")
-        result.to_csv(f"am/MARGOT/version_1/fin_{name}.xlsx")
+            # Run Margot over the documents
+            result = extract_argumentation()
+
+            # # Write result data/argumentation_structure
+            write_out(result, f"data/clean/test/annotated/{name}", cols=["text", "label"], tsv=True)
+
 
     # TODO: Train on the result
